JuaHaki-AI/requirements.txt
cohere
PyMuPDF
faiss-cpu
python-dotenv
numpy
sentence-transformers
google-generativeai
gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0